+++ Viewer/src/NodeQueryResult.cpp
@@ -161,7 +161,7 @@ NodeQueryResult::~NodeQueryResult()
 
 NodeQueryResultItem* NodeQueryResult::itemAt(int i)
 {
-	if(i >= 0 && i < data_.size())
+    if(i >= 0 && i < static_cast<int>(data_.size()))
 		return data_.at(i);
 
 	return NULL;
